Backend/requirements.txt
azure-storage-blob==12.19.1
bcrypt==4.0.1
celery==5.4.0
fastapi==0.110.2
#fitz==0.0.1.dev2
flower==2.0.1
gevent==24.2.1
llama-index==0.10.31
llama-index-llms-azure-openai==0.1.6
llama-index-embeddings-azure-openai==0.1.7
llama-index-vector-stores-postgres==0.1.5
motor==3.4.0
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic[email]
PyJWT==2.8.0
PyMuPDF==1.24.2
python-multipart==0.0.9
python-decouple==3.8
redis==5.0.4
SQLAlchemy==2.0.29
uvicorn==0.29.0
gunicorn==22.0.0